refactor(web): 컴포넌트 폴더 구조 개선 및 미사용 Header 제거 - 컴포넌트별 개별 폴더(index.tsx) 구조로 변경   (Footer, GNB, GlobalModal, GlobalToast, ThemeProvider, PlaceInfoCard) - 사용하지 않는 Header.tsx 삭제 - 관련 import 경로 수정
